almost complete migration to spring

+++ TopicHub-backend/src/main/java/com/example/topichubbackend/services/interfaces/IHubService.java
@@ -1,6 +1,7 @@
 package com.example.topichubbackend.services.interfaces;
 
 import com.example.topichubbackend.dto.*;
+import org.springframework.stereotype.*;
 
 import java.util.*;
 
@@ -9,7 +10,7 @@ public interface IHubService {
 
     HubDto create(HubDto hubDto);
 
-    void delete(Integer hubId);
+    void delete(Long hubId);
 
     HubDto update(HubDto hubDto);
 }
